zookeeper: join directly if --upgrade is set
This feature allows sheep to rejoin the cluster without sending a zk join event
after being killed as if the node were never dead. It take advantages of the zk
timeout window with temporal znode.

So the live upgrade is possible with this feature(rolling restart):

update sheep -> kill -9 sheep -> start sheep -old args -u within the zk timeout

WARN: We can't recover the block event states from the cluster, so the admin
should make sure there is NO BLOCK EVENT unhandled before rolling upgrade.

One idea to mitigate this problem is to check block events of the queue
carefully and if there IS block events unhandled, simply exit. I'd like to leave
this problem for the future with a (better) solution.

Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>